requirements.txt
django==3.1.3             # via -r requirements.in, djangorestframework, drf-yasg
djangorestframework==3.12.2  # via -r requirements.in, drf-yasg
gunicorn
django-heroku
django-cors-headers
